pyproject.toml
[tool.black]
line-length = 88
target_version = ['py38', 'py39', 'py310', "py311"]
exclude = '''
(
  /(
    | \.git
    | \.venv
    | \.mypy_cache
    | \.tox
  )/
)
'''

[tool.isort]
line_length = 88
profile = "black"

[tool.poetry]
name = "openshift-cli-installer"
version = "0.0.0"
description = "CLI to install/uninstall Openshift clusters."
readme = "README.md"
repository = "https://github.com/RedHatQE/openshift-cli-installer"
authors = ["Meni Yakove <myakove@gmail.com>", "Ruth Netser <rnetser@gmail.com>"]
license = "Apache-2.0"
packages = [{include = "openshift_cli_installer"}]
include = [{path = "manifests/*"}]
homepage = "https://github.com/RedHatQE/openshift-cli-installer"
documentation = "https://github.com/RedHatQE/openshift-cli-installer/blob/main/README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

[tool.poetry.scripts]
cli = "app.cli:main"

[tool.poetry.dependencies]
python = "^3.8"
jinja2 = "^3.1.2"
shortuuid = "^1.0.11"
click = "^8.1.4"
redhat-qe-cloud-tools = "^1.0.9"
rosa-python-client = "^1.0.16"
openshift-cluster-management-python-wrapper = "^1.0.27"
openshift-python-wrapper = "^4.14.1"
python-terraform = "^0.10.1"
semver = "^3.0.1"
pytest = "^7.4.0"
openshift-python-utilities = "^5.0.0"
pyaml-env = "^1.2.1"
google-cloud-compute = "^1.14.1"


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ipython = "*"

[tool.poetry-dynamic-versioning]
enable = true
pattern = "((?P<epoch>\\d+)!)?(?P<base>\\d+(\\.\\d+)*)"

[tool.poetry-dynamic-versioning.substitution]
files = ["VERSION"]

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning"]
build-backend = "poetry_dynamic_versioning.backend"
